# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BotAluguel Pro API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: List all users (admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /admin/stats:
    get:
      operationId: adminGetStats
      tags: [admin]
      summary: Get platform statistics (admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Platform stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
  /admin/payments:
    get:
      operationId: adminListPayments
      tags: [admin]
      summary: List all payments (admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
  /admin/notifications/send:
    post:
      operationId: adminSendNotification
      tags: [admin]
      summary: Send notification to users (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                body:
                  type: string
                type:
                  type: string
                targetUserId:
                  type: string
              required:
                - title
                - body
      responses:
        "200":
          description: Notification sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  sent:
                    type: number
                required:
                  - ok
                  - sent
  /admin/notifications:
    get:
      operationId: adminListNotifications
      tags: [admin]
      summary: List all notifications (admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AppNotification"
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List user notifications
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AppNotification"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        phone:
          type: string
        coins:
          type: number
        isAdmin:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - phone
        - coins
        - isAdmin
        - createdAt
...
    Payment:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        type:
          type: string
          enum: [topup, plan]
        amount:
          type: number
        coins:
          type: number
        status:
          type: string
          enum: [pending, paid, expired, error]
        txid:
          type: string
        createdAt:
          type: string
          format: date-time
        paidAt:
          type: string
          nullable: true
          format: date-time
      required:
        - id
        - userId
        - type
        - amount
        - coins
        - status
        - txid
        - createdAt
    AdminStats:
      type: object
      properties:
        totalUsers:
          type: number
        totalBots:
          type: number
        activeBots:
          type: number
        totalRevenue:
          type: number
        pendingPayments:
          type: number
        totalPlans:
          type: number
      required:
        - totalUsers
        - totalBots
        - activeBots
        - totalRevenue
        - pendingPayments
        - totalPlans
    AppNotification:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        title:
          type: string
        body:
          type: string
        type:
          type: string
        read:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - title
        - body
        - type
        - read
        - createdAt